Set delete_on_termination flag to false
When Terraform has to re-create a server, for example, due to the
user_data directive springing cloud-init into action, we do not want to
re-create the boot volume of the server. Hence, in the block_device
stanza of the openstack_compute_instance_v2 resource, we make sure to
set the delete_on_termination flag to false.